Oh no I'm not doing this again!

+++ src/lib.cpp
@@ -6,7 +6,7 @@ using namespace std;
 vector<char> split(const string toSplit, const string delimeter) {
 	vector<string> strings = splitStr(toSplit, delimeter);
 	vector<char> toReturn;
-	for (int i = 0; i < strings.size(); i++) {
+	for (int i = 0; i < (int)strings.size(); i++) {
 		if(!strings[i].empty()) toReturn.push_back(strings[i][0]);
 	}
 	return toReturn;
@@ -17,7 +17,7 @@ vector<string> splitStr(const string toSplit, const string delimeter) {
 	for (; indexAfter < toSplit.size() - 1;) {
 		size_t next = toSplit.find(delimeter, indexAfter);
 		string toAdd;
-		if (next == -1) {
+		if ((int)next == -1) {
 			toAdd = toSplit.substr(indexAfter, toSplit.size());
 			indexAfter = toSplit.size();
 		}
